Rename  folder to  for improved Next.js project structure
* Renamed the  folder to
* Aligns with **Next.js conventions** for project structure
* Improves **clarity and maintainability** of shared code
