+++ src/app/[locale]/(protected)/finance/_common/DeleteButton.js
