+++ src/app/[locale]/(protected)/finance/_common/list.jsx
