+++ src/app/[locale]/(protected)/finance/_common/view.module.css
