+++ src/app/[locale]/(protected)/finance/expenses/view/[id]/page.js
@@ -1,4 +1,4 @@
-import TransactionView from '../../../common/view'
+import TransactionView from '../../../_common/view'
 
 
 async function page({ params }) {
