+++ src/app/[locale]/(protected)/invoice/_common/DeleteButton.js
