+++ src/app/[locale]/(protected)/invoice/_common/RepositoryPermitButton.jsx
