+++ src/app/[locale]/(protected)/invoice/_common/view.js
