+++ src/app/[locale]/(protected)/invoice/purchases/list/page.js
@@ -1,5 +1,5 @@
 'use server'
-import InvoiceList from '../../common/list'
+import InvoiceList from '../../_common/list'
 
 async function page({ searchParams }) {
   return (
